requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
gspread>=6.0.0
google-auth>=2.28.0
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
lxml>=5.1.0
streamlit>=1.32.0
pandas>=2.2.0
